split cli.go into kubectl and mongocli
