+++ test/e2e/utils/mongocli.go
@@ -5,7 +5,6 @@ import (
 	"os/exec"
 
 	"encoding/json"
-	"strings"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -96,34 +95,5 @@ func GetClusterStatus(projectID string, clusterName string) func() string {
 	return func() string {
 		result := GetClustersInfo(projectID, clusterName)
 		return result.StateName
-	}
-}
-
-// GenKubeVersion TODO move
-func GenKubeVersion(fullVersion string) string {
-	version := strings.Split(fullVersion, ".")
-	return fmt.Sprintf("Major:\"%s\", Minor:\"%s\"", version[0], version[1])
-}
-
-// GetPodStatus TODO move
-func GetPodStatus(ns string) func() string {
-	return func() string {
-		session := Execute("kubectl", "get", "pods", "-l", "control-plane=controller-manager", "-o", "jsonpath={.items[0].status.phase}", "-n", ns)
-		return string(session.Wait("1m").Out.Contents())
-	}
-}
-
-func GetGeneration(nc string) func() string {
-	return func() string {
-		session := Execute("kubectl", "get", "atlascluster.atlas.mongodb.com/atlascluster-sample", "-n", nc, "-o", "jsonpath={.status.observedGeneration}")
-		return string(session.Wait("1m").Out.Contents())
-	}
-}
-
-// GetStatus TODO move
-func GetStatus(nc string, atlasname string) func() string {
-	return func() string {
-		session := Execute("kubectl", "get", atlasname, "-n", nc, "-o", "jsonpath={.status.conditions[?(@.type=='Ready')].status}")
-		return string(session.Wait("1m").Out.Contents())
 	}
 }
